Updated game for levels homework. Added custom sprites/assets to game, but still need to delete prof's nature-tiles.json.

+++ world/game_object.h
@@ -17,7 +17,7 @@ using Sprites = std::map<std::string, AnimatedSprite>;
 
 class GameObject {
 public:
-    GameObject(const Vec<int>& size, World& world, FSM* fsm, Input* input, Color color);
+    GameObject(const Vec<float>& size, World& world, FSM* fsm, Input* input, Color color);
     ~GameObject();
 
     void update(World& world, double dt);
@@ -27,7 +27,7 @@ public:
 
     // Player data
     Physics physics;
-    Vec<int> size;
+    Vec<float> size;
     FSM* fsm;
     Input* input;
     Color color;
